{{cookiecutter.project_slug}}/backend/app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Admin <admin@example.com>"]

[tool.poetry.dependencies]
python = "^3.7"
uvicorn = "^0.16.0"
fastapi = "^0.65.2"
python-multipart = "^0.0.5"
email-validator = "^1.0.5"
requests = "^2.23.0"
passlib = {extras = ["bcrypt"], version = "^1.7.2"}
tenacity = "^6.1.0"
pydantic = "^1.4"
emails = "^0.5.15"
raven = "^6.10.0"
gunicorn = "^20.0.4"
jinja2 = "^2.11.2"
psycopg2-binary = "^2.8.5"
alembic = "^1.4.2"
sqlalchemy = "^1.3.16"
python-jose = {extras = ["cryptography"], version = "^3.1.0"}
ariadne = "^0.14.0"
celery = "^5.2.3"
typer = "^0.4.0"
graphql-core = "<3.2.0"

[tool.poetry.dev-dependencies]
typed-ast = "^1.5.1"
mypy = "^0.930"
black = "^19.10b0"
isort = "^4.3.21"
autoflake = "^1.3.1"
flake8 = "^3.7.9"
sqlalchemy-stubs = "^0.3"
pytest-cov = "^2.8.1"
pytest = "^6.2.5"
pytest-dotenv = "^0.5.2"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 88
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

